Refactoring.
A few things have happened. First I renamed the parameter <n> to
<number> and instead I use <n> in the do-loop.

I've also added two lists of symbols and values (for X and V). Both
are used as stacks that I pop values from (into <s> and <v>) which
are used to identify which clause to execute.

In order to make this work I've changed from a cond-expression to
three when-expressions that are evaluated one after the other.
This allows us to translate, for example, 8 by first matching the
first when-expression (result = 'V', n = 3) and finally the last
one which yields result = 'VIII' and n = 0. A cond-expression had
only chosen one branch and then we would have popped another symbol
and value leaving us with an error since (>= n nil) doesn't work.

The last thing I did is that I modified the hndling of <n> when it
is 3 or less. The reason is the same as stated above. Unless I
handle it in one go, the do-loop while pop another value for <s>
and <v> leaving me with an error.