Refactoring tests.
Instead of having one test for each number I folded them into a
mapcar which returns a list of the values returned by the
function.

Next up is to add tests for up to thirty. Based on how this has
gone so far, I'd say we'll do fine up to 29 and then crash at 30.