Removed the test for 40.
This implementation is a dead-end. I'm going to have to refactor
the code before moving forward.

The pattern that is emerging is that numbers are translated based
on position with different symbols but the pattern stays the same.
So, for example, 8 (VIII) is very similar to 80 (LXXX) and 800
(DCCC).

The refactoring I'll try now is based around extracting each of
the digits in <number> and convert them one by one using different
symbols depending on if it's for 1s, 10s, 100s or 1000s.