Removed some tests and parts of the implementation.
Let's take a few steps back to (roughly) were the implementation
could manage all numbers up to 10. That's whre I started taking
shortcuts in the implementation.

Now that I know more about how this implementation is going to
turn out I can start the actual refactoring.

The first thing that needs to happen is to break this into two
functions. The function roman-numeral (as it is now) will be
doing the lion share of the work by producing roman numeral
patterns (like VII). The other function will split the number
into parts and provide the inner function with which symbols it
should fill the pattern with. XVI for up to ten, CLX for up to
100 and MDC for up to 1000.