Refactoring of convert-roman-numeral.
Ok. This will be the last polish on this function. It now returns
directly in case of 10, 9 and 4 and I've simplified the format
expressions.

The next stpe will be to try and tackle the larger numbers by
working on the outer function. I'll have to start by adding
support for up to 39 again.