Added functionality to solve the test case for 11.
I suspect that I've fixed quite a lot of the cases now. Probably
all the way up to 99. Fixing 100 and above is about adding another
line of roman numeral data.

Looking into the convert-roman-numeral function we can clearly see
that the when-clause checking for 10 will never happen so I'll do
some refactoring now.