Implementation.
Ok. This commit contains a little refactoring as well. I broke out
the same-suit implementaion and put in it's own function. At least
for now. That means I had to rename all the tests as well and that's
probably cheating.

The function same-suit-p takes any number of cards and returns t if
they're all in the same suit, otherwise it returns nil. It works by
looping through each of the cards, checking if they all belong to
(member) the same suit using the lists hearts and spades.