Implementation.
This implementation is fairly simple since most of the heavy
lifting is done in same-suit-p. I did however have to refactor
the same-suit-p function to work specifically with a list as
input argument instead of the implicit list as provided by the
&rest cards argument. The reason is that it is much easier for
me not to have to worry about splicing the arguments whenever
I need to call an underlying function.