Implementation.
Moved out the lists of suits to defparameters because I thought
I would need them in the implementation of high-card. However,
having finished that implementation now, it turned out I didn't.
I added another parameter instead called values.

The implementations is based around a sorted list of all cards
and each card has a value that is equal to the length of the list
returned by (member card *values*). The lower the better (at
least for now).