Implementation.
I reversed the list *values* so that the higher cards get a
higher score. Also updated the calculation of the score so
that it takes the length of the (member card *values) + 1 and
divides it by 4 and rounds upwards. That way. All the cards
of the same rank should get the same score regardless of which
suit they're in.