Refactoring.
Cleaned up a little bit in the documentation strings. Also
simplified two-pair-p to, essentially, two calls to pair-p and
some whens. The second call to pair-p is done with the cards
from the first pair taken out. I guess that is the trickiest
part of the code.

In order to achieve the simplification of two-pair-p I also
had to refactor the pair-p function to return two values. Both
the score and the symbols of the cards. Unfortunately pair-p
doesn't check that a pair actually consists of two cards which
is why (pair-p '(3S 3D 3C)) will return a list of three
symbols. I suspect that I will refactor the pair-p function
again when implementing scoring functions for 3 and 4 of a
kinds.