Implementation.
Added a modified version of pair-p in the three-of-a-kind
function. This logic should of course not be duplicated like
this but rather be moved into it's own function that take a
parameter for length so that I can use it for pair, 3 and 4
of a kind functions. I'll add some more tests (which I
expect to pass) and then do some refactoring (unless a test
fails of course).