Refactoring.
Moved out the implementation of three-of-a-kind into a
function called n-of-a-kind which takes two parameters, n and
hand. If it finds n cards of the same kind in hand it
returns the score and symbols of those cards. I also rewrote
the implementations for three-of-a-kind and pair-p to simply
be calls to n-of-a-kind with 2 and 3 respectively.