Test.
Added a function and a failing test for straight-p.

Looking at the body of functions that is growing into a set
of scoring functions I've started to think about a strategy
for the rank function. It seems like there's ranking done on
two levels. The first level is the rank (a flush beats a
straight etc) but the second is the ranking done if both
hands have the same rank. The general idea is that each of
the scoring functions will provide return values for both of
these rankings. That means I'll have to refactor some of the
functions (at least flush-p).