Implementation.
Added code for the straight-p function. The idea is to loop
through a sorted list of scored cards and make sure that they
only differ by 1 and otherwise return nil. Not particularly
happy with the code but can't put my finger on it, it's small
enough and not very complicated but ...