Test.
Added a test for the rank function (finally).

I've decided to have it return a list of scores that should be
compared, from left to right, until one player has a higher
score. If both lists are equal it's a tie.

The first score is based upon whether or not you have a pair,
three-of-a-kind etc. High card gives you 1 point, a pair 2,
etc up until a straight flush which gives you 9 points.

The rest of the scores are, for example, the score of the
cards that form your pair (in case all players have a pair).
All score lists contains all cards so that we only tie if
you have the same hand, just different suits.