Implementation.
Added a first cond for pairs. This should be fairly easy to
implement since most of the work is done elsewhere. However,
it's becoming clearer and clearer that all of those remove-if
calls should be refactored and placed in helper functions.

Also fixed a bug in full-house-p (there was a call to n-of-
a-kind left).