Implementation.
This was slightly trickier than I thought it would be. Turns
out I had to refactor two-pair-p somewhat. It now returns a
list of cards in the two pairs. Not a list of lists of cards.
Also, composing the scores is a bit different from the other
ones. Still hopeful it will be similar enough for some common
code though.